+++ src/org/uecide/varcmd/vc_lcase.java
@@ -1,0 +1,9 @@
+package org.uecide.varcmd;
+
+import org.uecide.*;
+
+public class vc_lcase extends VariableCommand {
+    public String main(Context sketch, String args) throws VariableCommandException {
+        return args.toLowerCase();
+    }
+}