+++ src/org/uecide/varcmd/vc_ucase.java
@@ -1,0 +1,9 @@
+package org.uecide.varcmd;
+
+import org.uecide.*;
+
+public class vc_ucase extends VariableCommand {
+    public String main(Context sketch, String args) throws VariableCommandException {
+        return args.toUpperCase();
+    }
+}